refactor: split module loaders logic
